// client/src/utils/consts.ts
export enum RouterNames {
  LOGIN_ROUTE = "/login",
  REGISTRATION_ROUTE = "/registration",
  START_PAGE = "/app",
  NODE = "/",
  USERS_ROUTE = "/app/users",
  VENDORS_ROUTE = "/app/vendors",
  TYPES_ROUTE = "/app/type",
  DEVICES_ROUTE = "/app/devices",
  PROPERTY_ROUTE = "/app/property",
  VALUE_TABLE_ROUTE = "/add/valuetable",
}

export enum pagination {
  pageSize = 5,
}